Major rewrite
This commit consists of a rewrite of basically all of the serialization
and deserialization logic. The API and internals are based heavily off
of serde_json, in particular, the error type and support for borrowed
deserialization. The deserialization logic is now based off of the tag
byte lookup table described in appendix B of RFC 7049. It's a bit more
verbose, but compiles down to a jump table so there should be some
decent performance benefits in deserialization. It should also resolve
some longstanding correctness issues with deserialization of indefinite
length maps and sequences. The packed encoding can now properly handle
conditionally skipped fields and doesn't need special logic on the
deserialization side due to new serde features.

Closes #37
Closes #36
Closes #32
Closes #31