Don't pass Visitor to parse_indefinite_{bytes,str}
It was used to create a `V::Value`. Because that was the last thing the
methods did, we can move the creation to the call sites instead.
Before this change, Rustc would compile a new version of these methods
for each Visitor.

This should improve compile times of crates using serde_cbor.