de/Read: Move scratch buffers to readers
This simplifies the parse_indefinite_* functions as they can now rely on
the full string to be in the reader's scratch buffer.

It adds a bit of interface to the Reader as the deserializer can now
specifically ask for reads to go directly to the scratch buffer or
(basically as before) to "simply read" (to scratch or a reference
return), plus the new function for clearing and accessing the buffer --
but then again the functions themselves are more straightforward (no
passing of a buf that may or may not be ignored by implementations).

The convert_str method was changed into an associated function; the
offset position is extracted from self and manually passed in to
convert_str. This is not strictly necessary at this point yet (given
that it's currenlty fed view_buffer results), but will be later.