Read API: Hand out variable-lifetime reads rather than deferring use to the buffer
Rather than having a Copied-or-with-Long-Liftetime Reference result
where Copied means that the user needs to go to the buffer for a
shortlived reference, deal out a variable-lifetime reference right away.

This helps keeping the scratch buffer more of a local thing, and opens
up the possibility for further optimization in the use of MutSliceRead.