MutSliceRead: Implement in a safe way
This is possible since there is no requirement to get an immutable view
on the scratch buffer as it is being built. The implementation now does
not use any bespoke unsafe code any more; the assertion on its internal
invariant is therefore only left in for debugging purposes as it's not
critical to memory safety any more.

Thanks to darkfox for pointing out the mem::replace(..., &mut []) trick
in [1].

[1]: https://gitlab.com/chrysn/sealingslice/issues/1