fixup serde_bytes dependency
When a crate depends on serde_cbor master, the following error is
thrown:
```
error: failed to select a version for `serde_cbor`.
    ... required by package `authenticator v0.2.4 (/home/baloo/work/dev/authenticator-rs)`
versions that meet the requirements `*` are: 0.9.0

the package `authenticator` depends on `serde_cbor`, with features: `serde_bytes` but `serde_cbor` does not have these features.

failed to select a version for `serde_cbor` which could resolve this conflict
```

I believe serde_bytes should not be a dev-dependency but an optional
dependency instead.

Signed-off-by: Arthur Gautier <baloo@gandi.net>