Don't abort build when clippy fails
Before, if the `clippy` check would fail `exit 0` would be called which
exited the test script. This meant that all subsequent steps would be
skipped and the test would pass.

Now instead of exiting the test script we just make sure that the script
always return a zero exit code.