Add tags support behind feature flag
Tags are stored as Newtype Structs with the special name `_TagStruct`. The tags
are represented as a struct tuple consisting of the tag identifier and a value.

Example:

    let tag = Value::Tag(123, Box::new(Value::Text("some value".to_string())));
    let tag_encoded = to_vec(&tag).unwrap();

This implementation is heavily based on the msgpack-rust library, when support
for extensions were added [1][2]. They have the same problem as CBOR has, that
such an extension/tagging type is not directly supported by Serde, hence a
workaround is needed. For me it makes sense to use a similarly working workaround
across different libraries.

This commit also contains an example on how to use custom types as tags. It can
be run via:

    cargo run --example tags --features tags

[1]: https://github.com/3Hren/msgpack-rust/commit/a34ab8fcca9f1133ec1c4fc446a59b71acf5613e
[2]: https://github.com/3Hren/msgpack-rust/pull/216