refactor: Refactor Values, MajorType and MinorType
This is yak shaving for OwnedValue, deserialization and streams.

Essentially, a Value now can contain references, but MajorType and
MinorType cannot. This is essential because when deserializing we
need to own the buffers that the Value references.

For example, if we wanted to create an OwnedValue (which owns its
own data), we would have had to re-create OwnedMajorType which was
a huge waste of efforts. Same when doing streams. Instead, now we
can reuse the full logic of MajorType and MinorType without worrying
about references and lifetime (because they don't have any anymore).

Please note that we did not change any tests, so we know the code
is compatible with our API.