Allow integer map keys to accept string-encoded integers
When serde attempts to deserialise a HashMap<i32, T> for example, with serde_json it will accept keys that are string-encoded versions of integers. This is because JSON only supports string keys. However this means that when you save data using serde_json, convert it to CBOR and then try and read it using serde_cbor, it will fail because CBOR does not accept string-encoded integer keys.

This commit changes that so that it does.