add RawValue for deferring/omitting serde-ing
Unfortunately this requires some unsafe, but fortunately it's only two
lines and it's very easy to follow. There are some problems still
though, such as not being able to deserialize to a `&RawValue` and
`Box<RawValue>`'s deserialize impl doing a roundtrip through `Value`.
`RawValue` could probably also be made available on alloc-only no-std
platforms, but I'm leaving it at std-only for now just in case I'm
wrong.

That aside, it does what it's supposed to do, so that's nice.